% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_cpm.R
\name{cpm_normalization}
\alias{cpm_normalization}
\title{Counts Per Million Normalization}
\usage{
cpm_normalization(x, log_trans = FALSE)
}
\arguments{
\item{log_trans}{Logical, whether to apply log2 transformation.}

\item{object}{A matrix, data.frame, list of data.frames, or a SummarizedExperiment object.}
}
\value{
Normalized object of the same structure.
}
\description{
Counts Per Million Normalization
}
